refactor: 애플리케이션 구조와 데스크톱 UX 정리
