--- src/features/Apps/AboutSite/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AppAboutSite';
