--- src/features/Apps/Config/allApps.tsx
@@ -1,7 +1,0 @@
-import DesktopApps from "./apps";
-import ProjectsApps from "./projectsApps";
-import { Language } from "@/shared/lib/i18n/useLanguageStore";
-
-export default function allApps(language: Language) {
-    return [...DesktopApps(language), ...ProjectsApps(language)];
-}
