--- src/features/Apps/Contact/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AppContact';
