--- src/features/Apps/DiceRoller/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AppDiceRoller';
