--- src/features/Apps/Flare/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AppFlare';
