--- src/features/Apps/Folder/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AppFolder';
