--- src/features/Apps/Mejubot/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AppMejubot';
