--- src/features/Apps/PageSsence/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AppPageSsence';
