--- src/features/Apps/Settings/index.ts
@@ -1,1 +1,0 @@
-export { default as LanguageSelector } from './LanguageSelector';
