--- src/features/Apps/Terminal/index.ts
@@ -1,3 +1,0 @@
-export { default as AppTerminal } from './AppTerminal';
-export { default as TerminalPage } from './TerminalPage';
-export * from './hooks/useTerminal';
