--- src/features/Apps/WCHMS/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AppWCHMS';
