--- src/features/Apps/WeConnect/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AppWeConnect';
