--- src/features/Apps/WebPiano/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AppWebPiano';
