+++ src/features/applications/components/flare/AppFlare.tsx
