+++ src/features/applications/components/terminal/AppTerminal.tsx
@@ -1,0 +1,14 @@
+'use client';
+
+import TerminalPage from "./TerminalPage";
+
+export default function AppTerminal() {
+
+    return (
+        <div
+            className="w-full h-full overflow-hidden bg-black rounded-b-pen-lg"
+        >
+            <TerminalPage />
+        </div>
+    );
+}
