+++ src/features/applications/components/terminal/TerminalRouteClient.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import dynamic from "next/dynamic";
+
+const TerminalPage = dynamic(() => import("./TerminalPage"), {
+    ssr: false,
+});
+
+export default function TerminalRouteClient() {
+    return <TerminalPage />;
+}
