+++ src/features/applications/components/terminal/command.ts
