+++ src/features/applications/components/terminal/lib/ascii.ts
