+++ src/features/applications/components/terminal/lib/bootSequence.ts
@@ -1,0 +1,16 @@
+export const TERMINAL_PROMPT = " guest@noonnofus.com ~ % ";
+
+export const TERMINAL_BOOT_LINES = [
+  " Hello, I'm Kevin, Kim.",
+  " I'm a full stack web developer.",
+  " Welcome to my portfolio.",
+  "    . ",
+  "    . ",
+  "    . ",
+  ' To see available commands, type "help" and press Enter.',
+];
+
+export const TERMINAL_BOOT_TIMING = {
+  asciiLine: 200,
+  character: 35,
+} as const;
