+++ src/features/applications/components/terminal/lib/commandParser.test.ts
