+++ src/features/applications/components/terminal/lib/commandParser.ts
