+++ src/features/applications/types/appDefinition.ts
@@ -1,0 +1,8 @@
+import type { ReactNode } from "react";
+
+export interface AppDefinition {
+    iconSrc: string;
+    appName: string;
+    title: string;
+    render: () => ReactNode;
+}
